Create IoT_final.py
This is the code for SensEve control system. Basically it is a IoT control weather system for home. It senses the weather outside using temperature and humidity sensor, and alters the temperature of the house using relay switches connect to HVAC system (Heating, Ventilation and Air Conditioning system). The system uses ultrasonic sensor to determine the occupancy so that, the intensity of the light changes accordingly to save power.